Seperated out route handler and helper functions for readability
