+++ server/src/routes/handlers/api/getCurrentScore.ts
@@ -1,0 +1,3 @@
+import { Request, Response } from "express";
+
+export const getCurrentScore = (req: Request, res: Response): void => {};
