+++ server/src/routes/handlers/api/index.ts
@@ -1,0 +1,4 @@
+export { getCurrentScore } from "./getCurrentScore";
+export { resetGame } from "./resetGame";
+export { startGame } from "./startGame";
+export { submitAnswer } from "./submitAnswer";
