+++ server/src/routes/handlers/api/resetGame.ts
@@ -1,0 +1,3 @@
+import { Request, Response } from "express";
+
+export const resetGame = (req: Request, res: Response): void => {};
