+++ server/src/routes/handlers/api/submitAnswer.ts
@@ -1,0 +1,3 @@
+import { Request, Response } from "express";
+
+export const submitAnswer = (req: Request, res: Response): void => {};
