+++ server/src/routes/helpers/buildRounds.ts
@@ -1,55 +1,4 @@
-import {
-  CharacterList,
-  CharacterResponse,
-  Round,
-  EmptyRound,
-  Character,
-  UserData,
-} from "../types";
-import https from "https";
-
-export const getCharacters = async (): Promise<CharacterList | Error> => {
-  return new Promise(async (resolve, reject) => {
-    try {
-      let characters: CharacterList = [];
-      let firstPage: CharacterResponse = (await getPage()) as CharacterResponse;
-      characters = characters.concat(firstPage.data);
-      let totalPages = firstPage.count;
-      let secondPage: CharacterResponse = (await getPage(
-        Math.floor(Math.random() * totalPages)
-      )) as CharacterResponse;
-      characters = characters.concat(secondPage.data);
-      resolve(characters);
-    } catch (err) {
-      reject(err as Error);
-    }
-  });
-};
-
-const getPage = (
-  pageNum: number = undefined
-): Promise<CharacterResponse | Error> => {
-  return new Promise((resolve, reject) => {
-    let pageString = "";
-    try {
-      https.get(
-        `https://api.disneyapi.dev/characters${
-          pageNum ? `?page=${pageNum}` : ""
-        }`,
-        (res) => {
-          res.on("data", (data) => {
-            pageString += data;
-          });
-          res.on("end", () => {
-            resolve(JSON.parse(pageString));
-          });
-        }
-      );
-    } catch (err) {
-      reject(err as Error);
-    }
-  });
-};
+import { CharacterList, Round, EmptyRound, Character } from "../../types";
 
 export const buildRounds = (characters: CharacterList): Round[] => {
   let validCharacters = characters.filter(
@@ -169,26 +118,4 @@ const getUniqueArray = (
       arrayToReadFrom
     );
   }
-};
-
-export const createUserGameData = async (
-  uuid: string
-): Promise<UserData | Error> => {
-  return new Promise(async (resolve, reject) => {
-    try {
-      const characters = await getCharacters();
-      if (characters) {
-        let rounds = buildRounds(characters as CharacterList);
-        resolve({
-          [uuid]: {
-            score: 0,
-            round: 0,
-            rounds: rounds,
-          },
-        } as UserData);
-      }
-    } catch (err) {
-      reject(err as Error);
-    }
-  });
 };
